Leverage ChainId inline class in more places

+++ rpc/src/main/kotlin/org/kethereum/rpc/Converter.kt
@@ -14,7 +14,7 @@ internal fun TransactionRPC.toKethereumTransaction() = SignedTransaction(
                 value = value.hexToBigInteger(),
                 from = Address(from),
                 to = to?.let { Address(it) },
-                chain = chainId?.hexToBigInteger()?.let { ChainId(it.toLong()) },
+                chain = chainId?.hexToBigInteger()?.let { ChainId(it) },
 
                 nonce = nonce.hexToBigInteger(),
                 gasPrice = gasPrice.hexToBigInteger(),
